+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/AndroidComponentsConfig.kt
@@ -98,7 +98,8 @@ fun AndroidComponentsExtension<*, *, *>.configure(
                     project,
                     extension.tracingInstrumentation.features,
                     extension.tracingInstrumentation.logcat.enabled,
-                    extension.includeSourceContext
+                    extension.includeSourceContext,
+                    extension.dexguardEnabled
                 )
                 /**
                  * We have to register SentryModulesService as a build event listener, so it will
@@ -274,7 +275,7 @@ private fun Variant.configureProguardMappingsTasks(
     if (isAGP74) {
         val variant = AndroidVariant74(this)
         val sentryProps = getPropertiesFilePath(project, variant)
-        val guardsquareEnabled = extension.experimentalGuardsquareSupport.get()
+        val guardsquareEnabled = extension.dexguardEnabled.get()
         val isMinifyEnabled = isMinificationEnabled(project, variant, guardsquareEnabled)
 
         if (isMinifyEnabled && extension.includeProguardMapping.get()) {